+++ ferricstore-spring-statemachine/src/main/java/com/ferricstore/spring/statemachine/DeniedTransitionPolicy.java
@@ -1,0 +1,7 @@
+package com.ferricstore.spring.statemachine;
+
+public enum DeniedTransitionPolicy {
+    FAIL,
+    RETRY,
+    THROW
+}